\name{Sample Data}
\alias{sample.data}
\docType{data}
\title{Sample dataset}

\description{
      Subset of BPG microarray dataset.
}

\usage{sample.data}

\format{
      A data frame with 5 columns and 5 rows. The columns indicate the patient sample,
      and the rows indicate the gene.
}

\author{Boutros Lab}

\keyword{datasets}
